requirements.txt
albumentations
matplotlib
numpy
opencv-python
Pillow
scikit-learn
seaborn
streamlit
tensorflow